+++ packages/website/__fixtures__/custom-webpack-project/docs.config.js
@@ -8,7 +8,7 @@ const webpack = config => {
 
 module.exports = () => ({
   packages: path.join(__dirname, 'packages', '*'),
-  docs: path.join(__dirname, 'docs'),
+  docs: { path: path.join(__dirname, 'docs'), name: 'docs' },
   siteName: 'Webpack Test Docs',
   webpack,
 });