+++ packages/website/__fixtures__/default-config/docs.config.js
@@ -1,4 +1,4 @@
 module.exports = () => ({
-  docs: 'a/b/c',
+  docs: { path: 'a/b/c', name: 'docs' },
   packages: ['x/y/z'],
 });