+++ packages/website/pages/guides/index.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import Wrapper from '../../src/components/page-templates/documents-index';
+
+export default () => (
+  <Wrapper data={{ key: 'guides', pageTitle: 'guides', pageType: 'docs' }} />
+);