+++ packages/website/src/bin/handle-config.test.js
@@ -9,18 +9,30 @@ describe('website configuration processor', () => {
 
     expect(config).toEqual({
       packagesPaths: ['/c/w/d/packages/*'],
-      docsPath: '/c/w/d/docs',
+      docsList: [
+        {
+          description: 'View the documentation for this project',
+          docsPath: '/c/w/d/docs',
+          name: 'Docs',
+        },
+      ],
       useManifests: false,
       siteName: 'Brisk Docs',
       webpack: expect.any(Function),
       showSubExamples: false,
+      packagesDescription: 'View documentation about individual packages',
     });
   });
 
   it('accepts a docs path as a string', () => {
-    const config = processConfig(mockCwd, { docs: 'some/other/docs' });
-
-    expect(config.docsPath).toEqual('/c/w/d/some/other/docs');
+    const { docsList } = processConfig(mockCwd, {
+      docs: {
+        path: 'some/other/docs',
+        name: 'Docs',
+        description: 'View the documentation for this project',
+      },
+    });
+    expect(docsList[0].docsPath).toEqual('/c/w/d/some/other/docs');
   });
 
   it('accepts an packages path as a string', () => {
@@ -63,17 +75,38 @@ describe('loadConfig', () => {
 
   it('should load a config given a valid config path', () => {
     expect(loadConfig(defaultConfigPath, 'custom-config-file.js')).toEqual({
-      docs: 'now/is/the/winter',
+      docs: {
+        path: 'now/is/the/winter',
+        description: 'View custom documentation',
+      },
       packages: ['of/our/disco/tents'],
+      packagesDescription:
+        'View custom documentation about individual packages',
     });
   });
   it('should load the default config if no path is given', () => {
     expect(loadConfig(defaultConfigPath)).toEqual({
-      docs: 'a/b/c',
+      docs: { path: 'a/b/c', name: 'docs' },
       packages: ['x/y/z'],
     });
   });
-  it('should return an empty object if there is no default', () => {
-    expect(loadConfig(__dirname)).toEqual({});
+  it('should load all the items in the docs as array in a config', () => {
+    expect(loadConfig(defaultConfigPath, 'docs-with-array-config.js')).toEqual({
+      docs: [
+        {
+          path: 'now/is/the/winter',
+          name: 'docs',
+          description: 'View custom documentation',
+        },
+        {
+          path: 'now/is/the/summer',
+          name: 'guides',
+          description: 'View custom guides',
+        },
+      ],
+      packages: ['of/our/disco/tents'],
+      packagesDescription:
+        'View custom documentation about individual packages',
+    });
   });
 });