+++ packages/website/src/bin/page-generator/write-pages.js
@@ -1,4 +1,4 @@
-const fse = require('fs-extra');
+const fs = require('fs-extra');
 const path = require('path');
 
 const {
@@ -10,8 +10,8 @@ const {
 } = require('./templates');
 
 const writeFile = (pagePath, content) => {
-  fse.ensureFileSync(pagePath);
-  fse.writeFileSync(pagePath, content);
+  fs.ensureFileSync(pagePath);
+  fs.writeFileSync(pagePath, content);
 };
 
 /**
@@ -207,6 +207,17 @@ const generateDocsHomePage = (pagePath, data, config, title = '') => {
   generateNonComponentPage(pagePath, data, 'item-list', config, 'docs', title);
 };
 
+const generateDocumentsMainPage = (pagePath, data, config, title = '') => {
+  generateNonComponentPage(
+    pagePath,
+    data,
+    'documents-index',
+    config,
+    'docs',
+    title,
+  );
+};
+
 const generateExamplesHomePage = (pagePath, data, config, title = '') => {
   generateNonComponentPage(
     pagePath,
@@ -243,4 +254,5 @@ module.exports = {
   generateDocsHomePage,
   generateExamplesHomePage,
   generateProjectDocPage,
+  generateDocumentsMainPage,
 };