+++ packages/website/src/components/mdx/index.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import Heading from './Heading';
 import Link from '../switch-link';
+import HorizontalRule from './HorizontalRule';
 import { CodeBlock, InlineCode } from './code';
 
 const components = {
@@ -13,6 +14,7 @@ const components = {
   a: Link,
   code: CodeBlock,
   inlineCode: InlineCode,
+  hr: HorizontalRule,
 };
 
 export default components;