+++ packages/website/src/components/meta-context.tsx
@@ -2,9 +2,14 @@ import * as React from 'react';
 
 interface ContextInterface {
   siteName: string;
+  packagesDescription: string;
+  [key: string]: any;
 }
 
-const ContextMeta = React.createContext<ContextInterface>({ siteName: '' });
+const ContextMeta = React.createContext<ContextInterface>({
+  siteName: '',
+  packagesDescription: '',
+});
 
 export default ContextMeta;