+++ packages/website/src/components/navigation/docs-nav-content.tsx
@@ -1,30 +1,30 @@
 import * as React from 'react';
-import { BackItem, MenuSection, Separator } from '@atlaskit/navigation-next';
+import { MenuSection } from '@atlaskit/navigation-next';
 import pageInfo from '../../pages-list';
-import LinkComponent from './link-component';
 import NavHeader from './nav-header';
 import TreeNavContent, { arrayToTreeItems } from './tree-nav-content';
 
-const NavContent = () => {
+const NavContent = ({ docId }: { docId: string }) => {
   const treeData = {
-    rootId: 'docs',
-    items: pageInfo.docs
-      ? arrayToTreeItems(pageInfo.docs, {
-          parentId: 'docs',
-          parentTitle: 'Docs',
+    rootId: docId,
+    items: pageInfo[docId]
+      ? arrayToTreeItems(pageInfo[docId], {
+          parentId: docId,
+          parentTitle: docId,
         })
       : {},
   };
 
   return (
     <>
-      <NavHeader headerText="Docs" />
+      <NavHeader headerText={docId} url={`/${docId.toLowerCase()}`} />
       <MenuSection id="docs-section" parentId="index-section">
         {/* TODO: TSFix nav typing */}
         {({ className }: { className: string }) => (
           <div className={className}>
-            <BackItem text="Back to home" href="/" component={LinkComponent} />
-            <Separator />
+            {/* TODO: Hiding for the time being */}
+            {/* <BackItem text="Back to home" href="/" component={LinkComponent} /> */}
+            {/* <Separator /> */}
             <TreeNavContent treeData={treeData} />
           </div>
         )}