+++ packages/website/src/components/navigation/package-nav-content.tsx
@@ -1,14 +1,8 @@
 import * as React from 'react';
 
-import {
-  MenuSection,
-  BackItem,
-  Separator,
-  Group,
-} from '@atlaskit/navigation-next';
+import { MenuSection, Separator, Group } from '@atlaskit/navigation-next';
 import titleCase from 'title-case';
 import LinkWithRouter from './link-with-router';
-import LinkComponent from './link-component';
 import pageInfo from '../../pages-list';
 import NavHeader from './nav-header';
 import TreeNavContent, { arrayToTreeItems } from './tree-nav-content';
@@ -54,16 +48,10 @@ const NavContent = ({
   subExamples,
 }: SomeProps) => (
   <>
-    <NavHeader headerText={packageName} />
+    <NavHeader headerText={packageName} url="/packages" />
     <MenuSection id="package-section" parentId="index-section">
       {({ className }: { className: string }) => (
         <div className={className}>
-          <BackItem
-            text="Back to packages"
-            href="/packages"
-            component={LinkComponent}
-          />
-          <Separator />
           <LinkWithRouter text="Home" href={homePath} />
           <Separator />
           {changelogPath && (