+++ packages/website/src/components/navigation/tree-nav-content.tsx
@@ -1,11 +1,21 @@
 import * as React from 'react';
+import styled from '@emotion/styled';
 import { Item } from '@atlaskit/navigation-next';
 import Tree from '@atlaskit/tree';
 import titleCase from 'title-case';
+import { colors } from '@atlaskit/theme';
 import LinkWithRouter from './link-with-router';
 
+const ParentWrapper = styled.div`
+  font-size: 10px;
+  margin: 4px 0;
+  padding: 0;
+  color: ${colors.N200};
+  text-transform: uppercase;
+`;
+
 type Item =
-  | { id: string; children: Item[] }
+  | { id: string; pagePath: string; children: Item[] }
   | { id: string; pagePath: string; children: undefined };
 
 type ArrayItems = Item[];
@@ -13,16 +23,21 @@ type ArrayItems = Item[];
 // Flatten the nested page structure into an object that ak/tree understands
 const arrayToTreeItems = (
   arrayItems: ArrayItems,
-  { parentId, parentTitle }: { parentId: string; parentTitle: string },
+  {
+    parentId,
+    parentTitle,
+    parentPath,
+  }: { parentId: string; parentTitle: string; parentPath?: string },
 ): any => ({
   // TODO: TSFix - this is return type to be specific
   [parentId]: {
     id: parentId,
-    hasChildren: true,
+    isHeading: !!arrayItems.length,
     isExpanded: true,
     children: arrayItems.map(sub => `${parentId}/${sub.id}`),
     data: {
       title: parentTitle,
+      href: parentPath,
     },
   },
   ...arrayItems.reduce((acc, sub) => {
@@ -34,6 +49,7 @@ const arrayToTreeItems = (
         ...arrayToTreeItems(sub.children, {
           parentId: id,
           parentTitle: sub.id,
+          parentPath: sub.pagePath,
         }),
       };
     }
@@ -43,7 +59,6 @@ const arrayToTreeItems = (
       [id]: {
         id,
         children: [],
-        hasChildren: false,
         data: {
           title: sub.id,
           href: sub.pagePath,
@@ -56,6 +71,7 @@ const arrayToTreeItems = (
 type TreeItemProps = {
   item: {
     id: string;
+    isHeading: boolean;
     hasChildren: boolean;
     data: {
       href: string;
@@ -69,20 +85,25 @@ type TreeItemProps = {
 };
 
 const TreeItem = ({ item, provided }: TreeItemProps) => {
-  const { id, hasChildren, data } = item;
-
-  if (!hasChildren) {
+  const { id, data, isHeading } = item;
+  const text = titleCase(data.title);
+  if (data.href) {
     return (
       <div ref={provided.innerRef} {...provided.draggableProps}>
-        <LinkWithRouter href={data.href} id={id} text={titleCase(data.title)} />
+        <LinkWithRouter
+          isHeading={isHeading}
+          href={data.href}
+          id={id}
+          text={text}
+        />
       </div>
     );
   }
 
   return (
-    <div ref={provided.innerRef} {...provided.draggableProps}>
-      <Item id={id} text={titleCase(data.title)} />
-    </div>
+    <ParentWrapper ref={provided.innerRef} {...provided.draggableProps} id={id}>
+      {text}
+    </ParentWrapper>
   );
 };
 
@@ -94,7 +115,7 @@ export type Props = {
 };
 
 const TreeNavContent = ({ treeData }: Props) => (
-  <Tree tree={treeData} renderItem={TreeItem} />
+  <Tree offsetPerLevel={16} tree={treeData} renderItem={TreeItem} />
 );
 
 export { arrayToTreeItems };