+++ packages/changeset-bot/CHANGELOG.md
@@ -1,5 +1,18 @@
 # changeset-bot
 
+## 1.1.2
+### Patch Changes
+
+- a18dd18: Got feedback that our changeset messages could be a bit loud/scary. Here is some lighter, friendlier text
+
+## 1.1.1
+- [patch] fa74183:
+
+  - Link messages out to information on changesets
+- [patch] a1d1b56:
+
+  - Display commit sha instead of commit message in bot comment
+
 ## 1.1.0
 - [patch] f7a6395: